+++ geotrek/settings/base.py
@@ -795,6 +795,8 @@ env_settings_file = os.path.join(os.path.dirname(__file__), 'env_{}.py'.format(E
 with open(env_settings_file, 'r') as f:
     exec(f.read())
 
+COLUMNS_LISTS = {}
+
 # Override with custom settings
 custom_settings_file = os.getenv('CUSTOM_SETTINGS_FILE')
 if custom_settings_file and 'tests' not in ENV: