+++ geotrek/trekking/views.py
@@ -1,3 +1,4 @@
+
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.contrib.gis.db.models.functions import Transform
@@ -19,6 +20,7 @@ from rest_framework import permissions as rest_permissions, viewsets
 
 from geotrek.api.v2.functions import Length
 from geotrek.authent.decorators import same_structure_required
+from geotrek.common.mixins import CustomColumnsMixin
 from geotrek.common.models import Attachment, RecordSource, TargetPortal, Label
 from geotrek.common.views import (FormsetMixin, MetaMixin, DocumentPublic,
                                   DocumentBookletPublic, MarkupPublic)
@@ -59,10 +61,11 @@ class TrekLayer(MapEntityLayer):
     queryset = Trek.objects.existing()
 
 
-class TrekList(FlattenPicturesMixin, MapEntityList):
+class TrekList(CustomColumnsMixin, FlattenPicturesMixin, MapEntityList):
     filterform = TrekFilterSet
-    columns = ['id', 'name', 'duration', 'difficulty', 'departure', 'thumbnail']
     queryset = Trek.objects.existing()
+    mandatory_columns = ['id', 'name']
+    default_extra_columns = ['duration', 'difficulty', 'departure', 'thumbnail']
 
 
 class TrekJsonList(MapEntityJsonList, TrekList):
@@ -70,8 +73,9 @@ class TrekJsonList(MapEntityJsonList, TrekList):
 
 
 class TrekFormatList(MapEntityFormat, TrekList):
+    mandatory_columns = ['id']
     columns = [
-        'id', 'eid', 'eid2', 'structure', 'name', 'departure', 'arrival', 'duration',
+        'eid', 'eid2', 'structure', 'name', 'departure', 'arrival', 'duration',
         'duration_pretty', 'description', 'description_teaser',
         'networks', 'advice', 'ambiance', 'difficulty', 'information_desks',
         'themes', 'practice', 'accessibilities', 'access', 'route',
@@ -244,11 +248,12 @@ class POILayer(MapEntityLayer):
     properties = ['name', 'published']
 
 
-class POIList(FlattenPicturesMixin, MapEntityList):
+class POIList(CustomColumnsMixin, FlattenPicturesMixin, MapEntityList):
     model = POI
-    filterform = POIFilterSet
-    columns = ['id', 'name', 'type', 'thumbnail']
     queryset = model.objects.existing()
+    filterform = POIFilterSet
+    mandatory_columns = ['id', 'name']
+    default_extra_columns = ['type', 'thumbnail']
 
 
 class POIJsonList(MapEntityJsonList, POIList):
@@ -256,14 +261,13 @@ class POIJsonList(MapEntityJsonList, POIList):
 
 
 class POIFormatList(MapEntityFormat, POIList):
-    columns = [
+    mandatory_columns = ['id']
+    default_extra_columns = [
         'id', 'structure', 'eid', 'name', 'type', 'description', 'treks',
         'review', 'published', 'publication_date',
         'structure', 'date_insert', 'date_update',
         'cities', 'districts', 'areas'
     ] + AltimetryMixin.COLUMNS
-
-    set(POIList.columns + ['description', 'treks', 'districts', 'cities', 'areas', 'structure'])
 
     def get_queryset(self):
         qs = super().get_queryset()
@@ -430,9 +434,10 @@ class ServiceLayer(MapEntityLayer):
     queryset = Service.objects.existing()
 
 
-class ServiceList(MapEntityList):
+class ServiceList(CustomColumnsMixin, MapEntityList):
     filterform = ServiceFilterSet
-    columns = ['id', 'name']
+    mandatory_columns = ['id', 'name']
+    default_extra_columns = []
     queryset = Service.objects.existing()
 
 
@@ -441,7 +446,8 @@ class ServiceJsonList(MapEntityJsonList, ServiceList):
 
 
 class ServiceFormatList(MapEntityFormat, ServiceList):
-    columns = [
+    mandatory_columns = ['id']
+    default_extra_columns = [
         'id', 'eid', 'type'
     ] + AltimetryMixin.COLUMNS