+++ docs/changelog.rst
@@ -5,6 +5,15 @@ CHANGELOG
 2.61.1+dev (XXX-XX-XX)
 ----------------------
 
+**New features**
+
+- Add custom columns configuration to list views 
+- Add custom columns configuration to list CSV exports
+
+**Bug fixes**
+
+* Fix filter difficulty in API v2
+
 
 2.61.1 (2021-06-28)
 -------------------